feat:finalização de testes

+++ src/services/User/DestroyUserService.ts
@@ -1,5 +1,6 @@
 import { inject, injectable } from 'tsyringe';
 import { UserRepositoryInterface } from '../../repositories/UserRepositoryInterface';
+import { ADMIN } from '../../../constants/roles';
 
 interface DataUser {
     first_name: string;
@@ -14,7 +15,11 @@ class DestroyUserService {
 
     constructor(@inject('UserRepository') private userRepository: UserRepositoryInterface) {}
    
-    async execute(id: number): Promise<void> {
+    async execute(id: number, userLogged: any): Promise<void> {
+        if(userLogged.role_id != ADMIN || userLogged.id === +id) {
+            throw new Error("Usuário não possui acesso a este recurso");
+        }
+
         try {
             await this.userRepository.delete(id);
         }catch(error) {
